+++ fixtures/ext-types/transitive-lib/tests/test_generated_bindings.rs
@@ -1,0 +1,1 @@
+uniffi::build_foreign_language_testcases!("tests/bindings/test_transitive_external_types.rb");